Move child issue components
