+++ src/components/EpicDetailView/EpicDetailView.tsx
@@ -23,7 +23,7 @@ import { ReporterMenu } from "../DetailView/Components/ReporterMenu"
 import { DeleteIssue } from "../DetailView/Components/DeleteIssue"
 import { ColorSchemeToggle } from "../common/ColorSchemeToggle"
 import { IssueIcon } from "../BacklogView/Issue/IssueIcon"
-import { ChildIssues } from "./helpers/ChildIssues"
+import { ChildIssues } from "./Components/ChildIssue/ChildIssues"
 import { getBacklogIssues } from "../BacklogView/helpers/queryFetchers"
 import { sortIssuesByRank } from "../BacklogView/helpers/backlogHelpers"
 import { useCanvasStore } from "../../lib/Store"
